linked cluster test
Test a linked cluster of three Docker instances. This test case simulates a
situation where instance A can reach B using an address that only has meaning
to A. The same goes for B to C. This means that the address that A uses to connect
to B can not be used by C to connect to B. The address used by B to connect to C can
not be used by A to connect to C. A is unreachable to C and the other way around.

The only path is then:

.===.   .===.   .===.
| A |->-| B |->-| C |
'==='   '==='   '==='

This scenario tests whether starting from A, slaving B will create a reverse
route using CJDNS from B to A and that slaving C from B will create another
reverse route. C should then be able to connect to A using the two tunnels
and vice versa. This simulates a firewalled scenario where a host can connect
to another host which then can also connect to yet another host, one that is
not accessible by the first host.
